+++ NEWS.md
@@ -1,4 +1,10 @@
 
+# fixest 0.14.2
+
+## Bug fixes
+
+- fix bug in vcov(): following v0.14.0, the returned object was a matrix of class `fixest_vcov` -- which led to problem for matrix operations as it didn't inherit the matrix class. Now fixed, thanks to @strengejacke, #652.
+
 # fixest 0.14.1
 
 ## Bug fixes